fix(ci): split unit-scale fixtures out of diff-test-helpers.ts to clear the module-size ratchet
The previous commit's new combined-model fixture pushed
packages/cli/src/commands/diff-test-helpers.ts to 425 lines with no
allowlist row, failing the same "Check TypeScript module size ratchet"
step this branch's earlier push was meant to clear (confirmed on the
fresh CI run for this branch).

Move every unit-scale STEP fixture (unitScaledPropertyModel,
unitScaledCombinedModel, and their UNIT_SCALE_* constants) into a new
sibling module, diff-unit-scale-fixtures.ts, importing guid() from
diff-test-helpers.ts. diff-test-helpers.ts is back to 313 lines;
diff-unit-scale-fixtures.ts is 124. Both the CLI and MCP test suites
that consume these fixtures now import them from the new file.
